Extract TryGetSemanticVersion method into extensions class
