+++ src/GitVersion.Core.Tests/VersionCalculation/Strategies/VersionInBranchNameBaseVersionStrategyTests.cs
@@ -33,10 +33,8 @@ public class VersionInBranchNameBaseVersionStrategyTests : TestBase
         baseVersion.SemanticVersion.ToString().ShouldBe(expectedBaseVersion);
     }
 
-    //[TestCase("hotfix-2.0.0")]
     [TestCase("origin/hotfix-2.0.0")]
     [TestCase("remotes/origin/hotfix-2.0.0")]
-    //[TestCase("hotfix/2.0.0")]
     [TestCase("origin/hotfix/2.0.0")]
     [TestCase("remotes/origin/hotfix/2.0.0")]
     [TestCase("custom/JIRA-123")]
@@ -70,6 +68,10 @@ public class VersionInBranchNameBaseVersionStrategyTests : TestBase
     [TestCase("release/3.0.0", "3.0.0")]
     [TestCase("support/2.0.0-lts", "2.0.0")]
     [TestCase("support-3.0.0-lts", "3.0.0")]
+    [TestCase("hotfix/2.0.0", "2.0.0")]
+    [TestCase("hotfix-3.0.0", "3.0.0")]
+    [TestCase("hotfix/2.0.0-lts", "2.0.0")]
+    [TestCase("hotfix-3.0.0-lts", "3.0.0")]
     public void CanTakeVersionFromNameOfConfiguredReleaseBranch(string branchName, string expectedBaseVersion)
     {
         using var fixture = new EmptyRepositoryFixture();
