+++ src/GitVersion.Core/Git/ReferenceName.cs
@@ -1,6 +1,4 @@
 using System.Diagnostics.CodeAnalysis;
-using System.Diagnostics.Contracts;
-using System.Text.RegularExpressions;
 using GitVersion.Extensions;
 using GitVersion.Helpers;
 
@@ -106,39 +104,7 @@ public class ReferenceName : IEquatable<ReferenceName?>, IComparable<ReferenceNa
         return Friendly;
     }
 
-    private char GetBranchSeparator() => WithoutOrigin.Contains('/') || !WithoutOrigin.Contains('-') ? '/' : '-';
-
     private static bool IsPrefixedBy(string input, string prefix) => input.StartsWith(prefix, StringComparison.Ordinal);
 
     private static bool IsPrefixedBy(string input, string[] prefixes) => prefixes.Any(prefix => IsPrefixedBy(input, prefix));
-
-    public bool TryGetSemanticVersion([NotNullWhen(true)] out (SemanticVersion Value, string? Name) result,
-        Regex versionPatternRegex, string? labelPrefix, SemanticVersionFormat format)
-    {
-        result = default;
-
-        Contract.Assume(versionPatternRegex.ToString().StartsWith("^"));
-
-        int length = 0;
-        foreach (var branchPart in WithoutOrigin.Split(GetBranchSeparator()))
-        {
-            if (branchPart.IsNullOrEmpty()) return false;
-
-            var match = versionPatternRegex.NotNull().Match(branchPart);
-            if (match.Success)
-            {
-                var versionPart = match.Groups["version"].Value;
-                if (SemanticVersion.TryParse(versionPart, labelPrefix, out var semanticVersion, format))
-                {
-                    length += versionPart.Length;
-                    var name = WithoutOrigin[length..].Trim('-');
-                    result = new(semanticVersion, name.IsEmpty() ? null : name);
-                    return true;
-                }
-            }
-            length += branchPart.Length + 1;
-        }
-
-        return false;
-    }
 }
